refactor(discovery): replace internal DnsSd class with stateless functions
The internal DnsSd class held no state beyond the adapter reference that
IrohNode already owns, so the class, its cached `#dnsSd` field, and the lazy
`#dns()` accessor were ceremony around two stateless operations. Move the
generic DNS-SD engine into `dns-sd.ts` as `advertiseService(adapter, options)`
and `browseServices(adapter, options)`, mirroring the existing `fetch.ts` /
`serve.ts` function modules. IrohNode.advertise/browse now call these directly.

No public API change: node.advertise/browse/advertisePeer/browsePeers are
unchanged.

Refs #329
